requirements.txt
flask==3.1.1
langchain==0.3.26
langchain-core==0.3.24
langchain-community==0.3.26
langchain-text-splitters==0.2.1
langchain-groq==0.2.3
langchain-pinecone==0.2.8
langchain-huggingface==0.0.1
pinecone-client==5.2.1
python-dotenv==1.1.0
sentence-transformers==4.1.0
pypdf==5.6.1
ipywidgets==8.1.1
jupyter==1.0.0
openai==1.14.0
-e .